requirements.txt
altair==4.2.0
async-generator==1.10
attrs==22.1.0
bcrypt==4.0.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cssselect==1.1.0
cssutils==2.6.0
decorator==5.1.1
deta==1.1.0
elastic-transport==7.13.0
elasticsearch==7.13.0
entrypoints==0.4
extra-streamlit-components==0.1.55
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
Jinja2==3.1.2
jsonschema==4.16.0
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
premailer==3.10.0
protobuf==3.20.2
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
python-dotenv==0.21.0
PyJWT==2.5.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rich==12.5.1
selenium==4.4.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
streamlit==1.12.2
streamlit-authenticator==0.1.5
toml==0.10.2
toolz==0.12.0
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wsproto==1.2.0
zipp==3.8.1
